Update tests and code to generate arrow data
Edit: cherry-picked commit to include the changes from julian hyde's
changes.

+++ arrow/src/main/java/org/apache/calcite/adapter/arrow/ArrowTranslator.java
@@ -196,9 +196,9 @@ class ArrowTranslator {
       BigDecimal bigDecimalLiteral = (BigDecimal) literal;
       int scale = bigDecimalLiteral.scale();
       if (scale == 0) {
-        return "long";
+        return "integer";
       } else if (scale > 0) {
-        return "double";
+        return "float";
       }
     } else if (String.class.equals(literal.getClass())) {
       return "string";
